+++ plugin-agave/src/protobuf/encoding/entry.rs
@@ -15,7 +15,7 @@ impl<'a> Entry<'a> {
     }
 }
 
-impl<'a> prost::Message for Entry<'a> {
+impl prost::Message for Entry<'_> {
     fn encode_raw(&self, buf: &mut impl bytes::BufMut) {
         let index = self.entry.index as u64;
         let starting_transaction_index = self.entry.starting_transaction_index as u64;