+++ CHANGELOG.md
@@ -18,6 +18,20 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ## 2025-03-12
 
+- cli-v4.0.0
+- client-v3.0.0
+- filter-v3.0.0
+- plugin-agave-v3.0.0
+- proto-v3.0.0
+- richat-v3.0.0
+- shared-v3.0.0
+
+### Breaking
+
+- solana: upgrade to v2.2 ([#85](https://github.com/lamports-dev/richat/pull/85))
+
+## 2025-03-12
+
 - richat-cli-v3.0.0
 - richat-filter-v2.4.1